fast-fix

+++ utilities/custom_slash.py
@@ -5,4 +5,5 @@ import discord
 async def auto_delete_webhook(interaction: discord.Interaction, content: str, delete_after: int, is_ephemeral=True):
     msg_to_delete = await interaction.followup.send(content, ephemeral=is_ephemeral)
     await asyncio.sleep(delete_after)
-    await msg_to_delete.delete()
+    if msg_to_delete:
+        await msg_to_delete.delete()
